# api-docs.yaml
openapi: 3.0.3
info:
  title: Geocoding API
  description: |
    A fast and comprehensive REST API for US ZIP code geocoding, Ohio address lookup, and county boundary data.
    
    ## 🔐 Authentication Required
    
    Most endpoints require an API key. To get started:
    
    1. **Register**: Visit the web interface and create an account
    2. **Get API Key**: Create an API key in your dashboard
    3. **Authenticate**: Use either authentication method:
       - `Authorization: Bearer your-api-key` header, OR
       - `X-API-Key: your-api-key` header
    4. **Test**: Use the "Authorize" button below to set your API key for all requests
    
    ## ✨ Features
    - **Lightning Fast**: PostgreSQL + PostGIS backed with optimized spatial indexes
    - **Multi-Level Coverage**: ZIP codes, cities, addresses, and county boundaries
    - **Comprehensive Data**: Population, density, coordinates, and geographic boundaries
    - **Flexible Search**: Multiple search modes with proximity and filtering
    - **ZIP Code Fallback**: City-based ZIP code lookup for flexible address searching
    - **Rich Metadata**: Timezone, county weights, administrative data
    - **GeoJSON Support**: Ready-to-use geographic data for mapping applications
    - **Usage Tracking**: Monitor your API usage and billing
    
    ## 📊 Data Coverage
    
    ### **🏙️ City Data**
    - **31,000+ US Cities**: Nationwide city coverage with demographics
    - **ZIP Code Associations**: All ZIP codes for each city
...
servers:
...
paths:
...
  /geocode/{zipcode}:
    get:
      summary: Get ZIP Code Details
      description: |
        Retrieve comprehensive geographic and demographic information for a specific US ZIP code.
        
        Returns detailed location data including coordinates, population statistics, 
        administrative boundaries, and timezone information.
        
        **Authentication Required**: This endpoint requires a valid API key using either `X-API-Key` header or `Authorization: Bearer` format.
      operationId: getZipCode
      security:
        - ApiKeyAuth: []
      tags:
        - Geocoding
      parameters:
        - name: zipcode
          in: path
          required: true
          description: US ZIP code (5 or 9 digits)
          schema:
            type: string
            pattern: '^\d{5}(-\d{4})?$'
            example: "10001"
      responses:
        '200':
          description: ZIP code found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeocodeResponse'
              examples:
                manhattan_zip:
                  summary: Manhattan ZIP code example
                  value:
                    success: true
                    data:
                      zip_code: "10001"
                      city_name: "New York"
                      state_code: "NY"
                      state_name: "New York"
                      zcta: true
                      population: 21102.0
                      density: 35400.5
                      primary_county_code: "36061"
                      primary_county_name: "New York"
                      county_weights:
                        "36061": "100"
                      county_names: ["New York"]
                      county_codes: ["36061"]
                      imprecise: false
                      military: false
                      timezone: "America/New_York"
                      latitude: 40.75066
                      longitude: -73.99670
                    count: 1
        '400':
          description: Invalid ZIP code format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Invalid ZIP code format"
        '404':
          description: ZIP code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "ZIP code not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /search:
    get:
      summary: Search ZIP Codes by City
      description: |
        Search for ZIP codes by city name with optional state filtering.
        
        Supports partial city name matching and returns multiple results.
        Useful for autocomplete functionality or when exact ZIP code is unknown.
      operationId: searchZipCodes
      security:
        - ApiKeyAuth: []
      tags:
        - Search
      parameters:
        - name: city
          in: query
          required: true
          description: City name or partial city name to search for
          schema:
            type: string
            minLength: 2
            example: "Springfield"
        - name: state
          in: query
          required: false
          description: Two-letter state code to filter results
          schema:
            type: string
            pattern: '^[A-Z]{2}$'
            example: "IL"
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
            example: 25
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              examples:
                springfield_search:
                  summary: Springfield search results
                  value:
                    success: true
                    data:
                      - zip_code: "62701"
                        city_name: "Springfield"
                        state_code: "IL"
                        state_name: "Illinois"
                        latitude: 39.78172
                        longitude: -89.64371
                      - zip_code: "62702"
                        city_name: "Springfield"
                        state_code: "IL"
                        state_name: "Illinois"
                        latitude: 39.76009
                        longitude: -89.60733
                    count: 2
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /distance/{from}/{to}:
    get:
      summary: Calculate Distance Between ZIP Codes
      description: |
        Calculate the precise distance between two ZIP codes using the Haversine formula.
        
        Returns distance in both miles and kilometers for accurate geographic calculations.
        Perfect for logistics, delivery radius validation, and distance-based searches.
      operationId: calculateDistance
      security:
        - ApiKeyAuth: []
      tags:
        - Distance
      parameters:
        - name: from
          in: path
          required: true
          description: Starting ZIP code
          schema:
            type: string
            pattern: '^\d{5}(-\d{4})?$'
            example: "10001"
        - name: to
          in: path
          required: true
          description: Destination ZIP code
          schema:
            type: string
            pattern: '^\d{5}(-\d{4})?$'
            example: "90210"
      responses:
        '200':
          description: Distance calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistanceResponse'
              example:
                success: true
                data:
                  from_zip_code: "10001"
                  to_zip_code: "90210"
                  distance_miles: 2445.5
                  distance_km: 3936.2
                count: 1
        '400':
          description: Invalid ZIP code format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /nearby/{zipcode}:
    get:
      summary: Find Nearby ZIP Codes
      description: |
        Find all ZIP codes within a specified radius of a center ZIP code.
        
        Returns ZIP codes sorted by distance with complete location data.
        Ideal for location-based services, delivery zones, and proximity marketing.
      operationId: findNearbyZipCodes
      security:
        - ApiKeyAuth: []
      tags:
        - Distance
      parameters:
        - name: zipcode
          in: path
          required: true
          description: Center ZIP code to search around
          schema:
            type: string
            pattern: '^\d{5}(-\d{4})?$'
            example: "10001"
        - name: radius
          in: query
          required: false
          description: Search radius in miles
          schema:
            type: number
            format: double
            minimum: 0.1
            maximum: 100
            default: 1
            example: 5
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
            example: 25
      responses:
        '200':
          description: Nearby ZIP codes found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearbyResponse'
              examples:
                nearby_results:
                  summary: ZIP codes within 5 miles of Manhattan
                  value:
                    success: true
                    data:
                      - zip_code:
                          zip_code: "10002"
                          city_name: "New York"
                          state_code: "NY"
                          latitude: 40.71571
                          longitude: -73.98803
                        distance_miles: 0.8
                        distance_km: 1.3
                      - zip_code:
                          zip_code: "10003"
                          city_name: "New York"
                          state_code: "NY"
                          latitude: 40.73168
                          longitude: -73.98975
                        distance_miles: 1.2
                        distance_km: 1.9
                    count: 2
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /proximity/{center}/{target}:
    get:
      summary: Check ZIP Code Proximity
      description: |
        Check if a target ZIP code is within a specified radius of a center ZIP code.
        
        Returns boolean result with actual distance for precise proximity checking.
        Useful for validation, filtering, and business rule enforcement.
      operationId: checkZipCodeProximity
      security:
        - ApiKeyAuth: []
      tags:
        - Distance
      parameters:
        - name: center
          in: path
          required: true
          description: Center ZIP code for proximity check
          schema:
            type: string
            pattern: '^\d{5}(-\d{4})?$'
            example: "10001"
        - name: target
          in: path
          required: true
          description: Target ZIP code to check proximity for
          schema:
            type: string
            pattern: '^\d{5}(-\d{4})?$'
            example: "10002"
        - name: radius
          in: query
          required: false
          description: Radius in miles for proximity check
          schema:
            type: number
            format: double
            minimum: 0.1
            maximum: 100
            default: 1
            example: 1
      responses:
        '200':
          description: Proximity check completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProximityResponse'
              example:
                success: true
                data:
                  center_zip_code: "10001"
                  target_zip_code: "10002"
                  radius_miles: 1
                  is_within_radius: true
                  actual_distance_miles: 0.5
                  actual_distance_km: 0.8
                count: 1
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /addresses:
    get:
      summary: Search Ohio Addresses
      description: |
        Search for addresses within Ohio with flexible filtering options.
        
        Supports searching by address components, city, county, or proximity to coordinates.
        Perfect for address validation, autocomplete, and location-based services.
      operationId: searchOhioAddresses
      security:
        - ApiKeyAuth: []
      tags:
        - Ohio Addresses
      parameters:
        - name: query
          in: query
          required: false
          description: General search query (searches across address fields)
          schema:
            type: string
            example: "123 Main St"
        - name: house_number
          in: query
          required: false
          description: House number to search for
          schema:
            type: string
            example: "123"
        - name: street
          in: query
          required: false
          description: Street name to search for
          schema:
            type: string
            example: "Main St"
        - name: city
          in: query
          required: false
          description: City name to filter by
          schema:
            type: string
            example: "Columbus"
        - name: county
          in: query
          required: false
          description: County name to filter by
          schema:
            type: string
            example: "Franklin"
        - name: postcode
          in: query
          required: false
          description: ZIP code to filter by
          schema:
            type: string
            example: "43215"
        - name: lat
          in: query
          required: false
          description: Latitude for proximity search (requires lon and radius)
          schema:
            type: number
            format: double
            example: 39.9612
        - name: lon
          in: query
          required: false
          description: Longitude for proximity search (requires lat and radius)
          schema:
            type: number
            format: double
            example: -82.9988
        - name: radius
          in: query
          required: false
          description: Search radius in miles for proximity search
          schema:
            type: number
            format: double
            minimum: 0.1
            maximum: 50
            default: 1
            example: 5
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
            example: 50
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: sort
          in: query
          required: false
          description: Sort order for results
          schema:
            type: string
            enum: [distance, address, city, county]
            default: address
            example: distance
      responses:
        '200':
          description: Address search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressSearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /addresses/semantic:
    get:
      summary: Semantic Address Search
      description: |
        Perform intelligent semantic search across Ohio addresses using flexible token-based matching.
        
        This endpoint uses advanced fuzzy matching and relevance scoring to find addresses regardless 
        of the order you specify address components. Perfect for autocomplete, address suggestions, 
        and flexible address lookups where users might enter components in any order.
        
        **Flexible Order Examples:**
        - `q=123 main columbus` → Finds "123 Main St" in Columbus
        - `q=columbus main 123` → Same result (order doesn't matter!)
        - `q=main st franklin county` → Finds Main Street addresses in Franklin County
        - `q=franklin 123 main` → Finds "123 Main St" in Franklin County
        - `q=43215 main` → Finds Main Street addresses in ZIP 43215
        - `q=columbus` → Finds addresses in Columbus
        - `q=main` → Finds all addresses on streets containing "main"
      operationId: semanticSearchAddresses
      security:
        - ApiKeyAuth: []
      tags:
        - Ohio Addresses
      parameters:
        - name: q
          in: query
          required: true
          description: Search query (partial address, street, city, etc.)
          schema:
            type: string
            minLength: 2
            example: "123 main st"
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return (1-50)
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
            example: 5
      responses:
        '200':
          description: Semantic search completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OhioAddress'
                  count:
                    type: integer
                    description: Number of addresses returned
                    example: 5
                  query:
                    type: string
                    description: The search query that was executed
                    example: "123 main st"
              example:
                success: true
                data:
                  - id: 12345
                    house_number: "123"
                    street: "Main St"
                    city: "Columbus"
                    county: "Franklin"
                    postcode: "43215"
                    latitude: 39.9612
                    longitude: -82.9988
                count: 5
                query: "123 main st"
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /addresses/{id}:
    get:
      summary: Get Address Details
      description: |
        Retrieve detailed information for a specific Ohio address by its unique ID.
        
        Returns complete address information including coordinates and administrative data.
      operationId: getOhioAddress
      security:
        - ApiKeyAuth: []
      tags:
        - Ohio Addresses
      parameters:
        - name: id
          in: path
          required: true
          description: Unique address ID
          schema:
            type: integer
            example: 12345
      responses:
        '200':
          description: Address found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '404':
          description: Address not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /counties:
    get:
      summary: List Ohio Counties
      description: |
        Retrieve a list of all Ohio counties with basic information and address counts.
        
        Supports filtering by name, address count ranges, and pagination.
        Perfect for building county selectors and administrative dashboards.
      operationId: getCounties
      security:
        - ApiKeyAuth: []
      tags:
        - County Boundaries
      parameters:
        - name: name
          in: query
          required: false
          description: Filter counties by name (partial match)
          schema:
            type: string
            example: "Franklin"
        - name: min_addresses
          in: query
          required: false
          description: Minimum number of addresses per county
          schema:
            type: integer
            minimum: 0
            example: 10000
        - name: max_addresses
          in: query
          required: false
          description: Maximum number of addresses per county
          schema:
            type: integer
            minimum: 0
            example: 500000
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
            example: 25
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        '200':
          description: Counties retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountyListResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /counties/{name}:
    get:
      summary: Get County Details
      description: |
        Retrieve detailed information about a specific Ohio county.
        
        Returns county metadata, statistics, and administrative information.
      operationId: getCountyDetail
      security:
        - ApiKeyAuth: []
      tags:
        - County Boundaries
      parameters:
        - name: name
          in: path
          required: true
          description: County name (case-insensitive)
          schema:
            type: string
            example: "Franklin"
      responses:
        '200':
          description: County found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountyResponse'
        '404':
          description: County not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/counties:
    get:
      summary: Get County Statistics
      description: |
        **Admin endpoint** to retrieve comprehensive statistics about Ohio counties.
        
        Returns aggregated data including total counties, address counts, and distribution metrics.
      operationId: getCountyStats
      tags:
        - Admin
      responses:
        '200':
          description: County statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountyStatsResponse'
        '500':
          description: Failed to retrieve statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Supports two methods:
        
        **Method 1 - X-API-Key Header (Recommended):**
        ```
        X-API-Key: your-api-key-here
        ```
        
        **Method 2 - Authorization Bearer:**
        ```
        Authorization: Bearer your-api-key-here
        ```
        
        Get your API key by:
        1. Register/login at the web interface
        2. Create an API key in your dashboard
        3. Use either authentication method above
...